chore: refactiring, notification component has been removed
