+++ packages/client/src/app/redux/store/reducers/index.ts
@@ -1,3 +1,2 @@
 export * from './userSlice'
 export * from './gameSlice'
-export * from './notificationSlice'
