--- packages/client/src/components/notification/index.ts
@@ -1,1 +1,0 @@
-export * from './Notification'
