+++ packages/client/src/shared/types/notification.ts
@@ -2,5 +2,4 @@ export enum NotificationStatus {
   GRANTED = 'granted',
   DENIED = 'denied',
   DEFAULT = 'default',
-  INVALID = 'invalid',
 }
